swic-max-speed.sh: New script
The script is used to find a maximum value of exchange frequency.

The script consistently iterates over a sequence of frequency
multipliers and does specified number of exchanging cycles for
every value. If an error occures the script saves the status
of exchange(hang or fail) and proceeds to the next speed value.

Issue: #MCOM02SW-1365
Change-Id: Ic639cee00e2eb43a6d16adfaca947513b67a9d33